+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+Under development
+-----------------
+- Minimal score validation
+
 1.1.4
-=====
+-----------------
 - RU and ET locales
 
 1.1.3
-=====
+-----------------
 - Explicit recaptcha rendering (for Pjax support and multiple widgets per page)
 
 1.1.2
-=====
+-----------------
 - Yii::trace() instead of Yii::debug() for backward compatibility